pyproject.toml
[tool.poetry]
name = "menu-app"
version = "0.1.0"
description = ""
authors = ["Dmitrii Turetskikh <tureckihdmitriq@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.26.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
pydantic-settings = "^2.1.0"
alembic = "^1.13.1"
pytest-asyncio = "^0.23.3"
pytest-dotenv = "^0.5.2"
httpx = "^0.26.0"
redis = "^5.0.1"
pre-commit = "^3.6.0"
types-redis = "^4.6.0.20240106"
mypy-extensions = "^1.0.0"
openpyxl = "^3.1.2"
celery = {extras = ["rabbitmq"], version = "^5.3.6"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
